[refactor] 드롭다운메뉴 경로 변경
